langchain/devops_mlops_agent/requirements.txt
openai
streamlit
pandas
scikit-learn
python-dotenv
